data_quality/requirements.txt
DateTime==4.3
numpy==1.17.3
pandas==1.3.1
PyMySQL==0.9.3
schedule==0.6.0
DBUtils==1.3.0
django==2.2.10
#2.2.10 3.2.6
python-cronTab==2.5.1
sqlalchemy==1.4.22
pandas==1.3.1
uwsgi==2.0.15
gunicorn==20.1.0
requests==2.22.0
paramiko==2.7.2
schedule==0.6.0
django-extensions==3.1.3
django-werkzeug-debugger-runserver==0.3.1
django-sslserver==0.22
pyOpenSSL==20.0.1
mysqlclient==2.0.3
#MySQL-python==1.2.5
flask==0.3.1
#requests==2.26.0
#sqlalchemy==1.4.25
pytz==2021.1
#pymysql==1.0.2
apscheduler==3.8.0
#pip install sasl
#pip install thrift
#pip install thrift-sasl
#pip install PyHive
#thrift-0.15.0